requirements.txt
pyserial
pandas
